requirements.txt
python-dotenv
streamlit
langchain
langchain-nomic
langchain-ollama
faiss-cpu
beautifulsoup4
requests
plotly
umap-learn